Support ligand building from script
