+++ .gitignore
@@ -137,3 +137,5 @@ dmypy.json
 
 # data
 data
+*.ckpt
+*.csv
