+++ pandda_gemmi/cnn/model_build_config.yaml
@@ -1,0 +1,34 @@
+sample_size: 32
+sample_spacing: 0.5
+lr: 0.00011438351645738927
+wd: 0.004660900954194315
+fraction_background_replace: 0.8456736753906298
+xmap_radius: 4.1308903965340065
+max_x_blur: 2.820896567200296
+max_z_blur: 0.6891269656644818
+drop_rate: 0.7688958132818271
+planes_1: 16
+drop_1: 0.01720841750409113
+planes_2: 32
+drop_2: 0.05160404771554821
+planes_3: 64
+drop_3: 0.003800989203779803
+planes_4: 128
+drop_4: 0.05806019760101572
+planes_5: 256
+drop_5: 0.6291779900450685
+drop_atom_rate: 0.5348397956124168
+max_pos_atom_mask_radius: 2.693877997882939
+max_translate: 3.960735194773154
+max_x_noise: 0.276422691457226
+max_z_noise: 0.0685760724760862
+pos_resample_rate: 23
+p_flip: 0.6278772700882774
+z_mask_radius: 2.005754066726283
+z_cutoff: 1.7837797684839956
+blocks_1: 2
+blocks_2: 2
+blocks_3: 2
+blocks_4: 1
+grad_clip: 0.0177383347423051
+batch_size: 32
