+++ pandda_gemmi/cnn/model_event_config.yaml
@@ -1,0 +1,37 @@
+sample_size: 32
+sample_spacing: 0.5
+lr: 0.000190173872510559
+wd: 0.03294736010282053
+fraction_background_replace: 0.9554165076086426
+xmap_radius: 3.350471097783028
+max_x_blur: 0.813050856940392
+max_z_blur: 0.7712571256720352
+drop_rate: 0.6198762110797235
+planes_1: 16
+drop_1: 0.23671534989833767
+planes_2: 32
+drop_2: 0.08637261578792743
+planes_3: 64
+drop_3: 0.03772067922082744
+planes_4: 32
+drop_4: 0.39873226037845494
+planes_5: 256
+drop_5: 0.643401608502051
+drop_atom_rate: 0.5255239477882974
+max_pos_atom_mask_radius: 2.9913975194967746
+max_translate: 0.8553965637468237
+max_x_noise: 1.8818277363226692
+max_z_noise: 0.9827872948352333
+pos_resample_rate: 9
+p_flip: 0.4497857609250661
+z_mask_radius: 2.716437477309731
+z_cutoff: 1.943722672574652
+combo_layer: 32
+blocks_1: 2
+blocks_2: 2
+blocks_3: 1
+blocks_4: 2
+grad_clip: 0.006557248285092227
+batch_size: 16
+ligand: true
+minimum_event_score: 0.15
